move basic network
